split out some data providing logic
