+++ src/perspective.mts
@@ -1,7 +1,6 @@
-import { vCross, vDot, vScale, vAdd, vSub } from "./quaternion.mjs";
+import { vCross, vDot, vScale, vAdd, vSub, V3 } from "./quaternion.mjs";
 import { coneBackScale } from "./config.mjs";
 import { Atom } from "./atom.mjs";
-import { V3 } from "./primes.mjs";
 
 export let atomViewerForward = new Atom<V3>([0, 0, -1]);
